requirements.txt
openai
azure-cognitiveservices-speech
evdev
httpx[socks]